Refactor Track model to use Long for size and add formatted size utility methods

+++ shared/src/main/java/com/ontrek/shared/data/Track.kt
@@ -14,6 +14,6 @@ data class Track(
     val filename: String,
     val stats: TrackStats,
     val title: String,
-    val size: Double,
+    val size: Long,
     val upload_date: String
 )
